Úpravy (#23)

+++ TemperatureSensors.h
@@ -18,7 +18,7 @@ class TemperatureSensors
     void GetAcumulator4Temperature(uint8_t* temperature);
     void GetAcumulatorOutputTemperature(uint8_t* temperature);
     void GetReturnHeatingTemperature(uint8_t* temperature);
-    void GetCurrentHeatingTemperature(float* temperature);
+    void GetCurrentHeatingTemperature(uint8_t* temperature);
     void GetHeaterTemperature(uint8_t* temperature);
     void GetBoilerTemperature(uint8_t* temperature);
     void Init();
